requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
langchain==0.0.335
langchain-community==0.0.13
openai==1.2.4
chromadb==0.4.18
pydantic==2.4.2
jinja2==3.1.2
sqlalchemy==2.0.23
python-multipart==0.0.6
tiktoken==0.5.1
